+++ lib/flutter_dev_utils.dart
@@ -1,6 +1,6 @@
 library flutter_dev_utils;
 
-export 'package:flutter_dev_utils/src/try_catch_handler/sync_try_catch_handler.dart';
-export 'package:flutter_dev_utils/src/try_catch_handler/async_try_catch_handler.dart';
+export 'package:flutter_dev_utils/src/try_handler/sync_try_handler.dart';
+export 'package:flutter_dev_utils/src/try_handler/async_try_handler.dart';
 export 'package:flutter_dev_utils/src/caller_logger/caller_logger.dart';
 export 'package:flutter_dev_utils/src/caller_logger/type_filter.dart';