+++ lib/src/inline_help/help_error.dart
@@ -1,0 +1,23 @@
+/// Flutter Error Types:
+///
+/// * ArgumentError
+/// * AssertionError
+/// * AsyncError
+/// * CastError
+/// * ConcurrentModificationError
+/// * CyclicInitializationError
+/// * DriverError
+/// * FallThroughError
+/// * FlutterError
+/// * HttpProfileRequestError
+/// * JsonUnsupportedObjectError
+/// * NoSuchMethodError
+/// * NullThrownError
+/// * OutOfMemoryError
+/// * RemoteError
+/// * StackOverflowError
+/// * StateError
+/// * TypeError
+/// * UnimplementedError
+/// * UnsupportedError
+abstract class HelpError {}