+++ lib/src/try_handler/async_try_handler.dart
@@ -1,13 +1,15 @@
-import 'package:flutter_dev_utils/src/try_catch_handler/logger.dart';
-import 'package:flutter_dev_utils/src/try_catch_handler/catch_utils.dart';
+import 'package:flutter_dev_utils/src/try_handler/logger.dart';
+import 'package:flutter_dev_utils/src/try_handler/catch_utils.dart';
 
 /// Asynchronous try and catch handler to reduce boilerplate
 ///
 /// Should be called in a State file
-Future<dynamic> asyncTryCatchHandler(
-    {required Future<dynamic> Function() tryFunction,
-    Map<dynamic, Future<dynamic> Function(Object e)>? catchKnownExceptions,
-    Future<dynamic> Function()? catchUnknownExceptions}) async {
+Future<dynamic> asyncTryHandler({
+  required Future<dynamic> Function() tryFunction,
+  Map<dynamic, Future<dynamic> Function(Object e)>? catchKnownExceptions,
+  Future<dynamic> Function(Object e)? catchUnknownExceptions,
+  dynamic finallyReturn,
+}) async {
   //! Validate Catch Known
   if (catchKnownExceptions != null) {
     validateCatchKnownExceptions(catchKnownExceptions);
@@ -40,6 +42,11 @@ Future<dynamic> asyncTryCatchHandler(
   } catch (e, s) {
     //! Handle Unknown Errors and Exceptions
     utilsLogger.e('Caught unknown exception', e, s);
-    return catchUnknownExceptions?.call();
+    return catchUnknownExceptions?.call(e);
+  } finally {
+    //! Finally
+    // ignore: control_flow_in_finally
+    return finallyReturn;
+    // this lint gets called because finally blocks should return const value
   }
 }