+++ test/flutter_dev_utils_test.dart
@@ -1,7 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:flutter_dev_utils/flutter_dev_utils.dart';
-
 void main() {
   test('test_async_try_catch_handler', () {
     expect('a', 'a');