+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+## 0.0.1
+
+* Added Help classes
+* Added finally to tryCatchHandler, renamed to tryHandler
+* Return error/exception in tryHandler.catchUnknownExceptions
+
 ## 0.0.0
 
 * tryCatchHandler and CallerLogger
-